remove strange index shift
might have to revert this change. self.points does not get reduced in size in the original program, just recorded with self.status_points. The problem is that KNN can only get indices in the real reduced point set, and would have to convert them to the higher serial indices corresponding to the full self.points